Refactor my profile pages
