+++ packages/ui/src/app/components/AppPage.tsx
@@ -1,14 +1,11 @@
 import React, { ReactNode } from 'react'
 import styled from 'styled-components'
 
-import { Page } from '../../common/components/page/Page'
 import { PageContent } from '../../common/components/page/PageContent'
 import { PageHeader } from '../../common/components/page/PageHeader'
 import { PageTabs, TabProps } from '../../common/components/page/PageTabs'
 import { PageTitle } from '../../common/components/page/PageTitle'
 import { Breadcrumbs } from '../../common/components/page/Sidebar/Breadcrumbs/Breadcrumbs'
-
-import { SideBar } from './SideBar'
 
 interface AppPageProps {
   crumbs: { href: string; text: string }[]
@@ -18,19 +15,16 @@ interface AppPageProps {
 }
 
 export const AppPage = ({ children, crumbs, pageTitle, tabs }: AppPageProps) => (
-  <Page>
-    <SideBar />
-    <PageContent>
-      <Breadcrumbs crumbs={crumbs} />
-      <PageContainer>
-        <PageHeader>
-          <PageTitle>{pageTitle}</PageTitle>
-          <PageTabs tabs={tabs} />
-        </PageHeader>
-        {children}
-      </PageContainer>
-    </PageContent>
-  </Page>
+  <PageContent>
+    <Breadcrumbs crumbs={crumbs} />
+    <PageContainer>
+      <PageHeader>
+        <PageTitle>{pageTitle}</PageTitle>
+        <PageTabs tabs={tabs} />
+      </PageHeader>
+      {children}
+    </PageContainer>
+  </PageContent>
 )
 
 export const PageContainer = styled.div`
