--- packages/ui/src/app/pages/Profile/MyAccounts/helpers.ts
@@ -1,16 +1,0 @@
-import { SortKey } from '../../../../accounts/model/sortAccounts'
-
-export function setOrder(
-  key: SortKey,
-  sortBy: SortKey,
-  setSortBy: (k: SortKey) => void,
-  reversed: boolean,
-  setDescending: (d: boolean) => void
-) {
-  if (key === sortBy) {
-    setDescending(!reversed)
-  } else {
-    setDescending(key !== 'name')
-    setSortBy(key)
-  }
-}
