--- packages/ui/src/app/pages/Profile/MyAccounts/index.ts
@@ -1,1 +1,0 @@
-export * from './MyAccounts'
