--- packages/ui/src/app/pages/Profile/MyMemberships/helpers.ts
@@ -1,16 +1,0 @@
-import { SortKey } from '../../../../memberships/model/sortMemberships'
-
-export function setOrder(
-  key: SortKey,
-  sortBy: SortKey,
-  setSortBy: (k: SortKey) => void,
-  reversed: boolean,
-  setDescending: (d: boolean) => void
-) {
-  if (key === sortBy) {
-    setDescending(!reversed)
-  } else {
-    setDescending(key !== 'handle')
-    setSortBy(key)
-  }
-}
