--- packages/ui/src/app/pages/Profile/MyMemberships/index.ts
@@ -1,1 +1,0 @@
-export * from './MyMemberships'
